Handle family creation robustly
When a new, non-fork post is saved, no new family is created yet.
When a fork is saved, it is only associated with a family when
published. In that case, the family of its parent is adopted. If the
parent is not associated with a family yet, it must be the case that
the parent is neither a fork itself nor publicly forked before, i.e.
it is about to become the root of a new family. In that particular
case, a new family is created with the same title as the parent. Both
the parent and the fork are then associated with the newly created
family.